[WIP] Fixes engine texture loading - on project load still invalid

+++ turnip-engine/src/graphics/vulkan/factories/pipeline_factory.cpp
@@ -247,7 +247,7 @@ namespace tur::vulkan
 			vk::PipelineLayoutCreateInfo pipelineLayoutCreateInfo = {};
 			{
 				pipelineLayoutCreateInfo.flags = vk::PipelineLayoutCreateFlags();
-				pipelineLayoutCreateInfo.setLayoutCount = setLayouts.size();
+				pipelineLayoutCreateInfo.setLayoutCount = static_cast<u32>(setLayouts.size());
 				pipelineLayoutCreateInfo.pSetLayouts = setLayouts.data();
 				pipelineLayoutCreateInfo.pushConstantRangeCount = 0;
 				pipelineLayoutCreateInfo.pPushConstantRanges = nullptr;
